Fix critical bug when toggling mathjax
The regex rules for mathjax including identifing inline and display
equations have been migrated from rules that were determined at
instantation of a marked object to static rule sets. Rules sets can now
be swaped out as required dynamically during runtime.

Previously, if {mathjax: false} was passed as an option, this would
irreversably change the the rule sets, such that if marked.setOptions
was used subsequently to toggle mathjax for the same marked instance,
mathjax would not be correctly identified and rendered.

This bug was present to less of a degree in the original changes made by
@summivox. No one probably noticed as no one had need to toggle mathjax
functionality during runtime.